Make start time a little earlier and reshuffle
We need a bit more time at the start (to wait for teams to arrive,
play the welcome video, and give teams a chance to register before the
practice contest starts).

And we should play the welcome video before the teams register,
because the video contains information about the registration system.